requirements.txt
python-telegram-bot
pydantic-settings